Make listener not fail on decode errors and db errors
