+++ src/relay/api.rs
@@ -17,7 +17,6 @@ use prometheus::TextEncoder;
 use protobuf::Message as ProtoMessage;
 use std::fmt::Display;
 use std::pin::Pin;
-use std::str::FromStr;
 use std::sync::Arc;
 use tokio::sync::Mutex;
 
@@ -204,27 +203,9 @@ async fn federation<D: TwinDB, R: RateLimiter>(
     let envelope =
         Envelope::parse_from_bytes(&body).map_err(|err| HttpError::BadRequest(err.to_string()))?;
 
-    if let Some(relays) = &envelope.relays {
-        let mut twin = data
-            .twins
-            .get_twin(envelope.source.twin)
-            .await
-            .map_err(|err| HttpError::FailedToGetTwin(err.to_string()))?
-            .ok_or_else(|| HttpError::TwinNotFound(envelope.source.twin))?;
-        let envelope_relays = match RelayDomains::from_str(relays) {
-            Ok(r) => r,
-            Err(_) => return Err(HttpError::BadRequest("invalid relays".to_string())),
-        };
-        if let Some(twin_relays) = twin.relay {
-            if twin_relays != envelope_relays {
-                twin.relay = Some(envelope_relays);
-                data.twins
-                    .set_twin(twin)
-                    .await
-                    .map_err(|err| HttpError::FailedToSetTwin(err.to_string()))?;
-            }
-        }
-    }
+    update_cache_relays(&envelope, &data.twins)
+        .await
+        .map_err(|err| HttpError::FailedToSetTwin(err.to_string()))?;
     let dst: StreamID = (&envelope.destination).into();
     data.switch.send(&dst, &body).await?;
 
@@ -232,6 +213,23 @@ async fn federation<D: TwinDB, R: RateLimiter>(
         .status(http::StatusCode::ACCEPTED)
         .body(Body::empty())
         .map_err(HttpError::Http)
+}
+
+async fn update_cache_relays(envelope: &Envelope, twin_db: &impl TwinDB) -> Result<()> {
+    if envelope.relays.len() == 0 {
+        return Ok(());
+    }
+    let twin = twin_db
+        .get_twin(envelope.source.twin)
+        .await?
+        .ok_or_else(|| anyhow::Error::msg("unknown twin source"))?;
+    let envelope_relays = RelayDomains::new(&envelope.relays);
+    if let Some(twin_relays) = twin.relay.clone() {
+        if twin_relays != envelope_relays {
+            twin_db.set_twin(twin).await?;
+        }
+    }
+    Ok(())
 }
 
 type Writer = SplitSink<WebSocketStream<Upgraded>, Message>;
@@ -331,24 +329,7 @@ impl<M: Metrics, D: TwinDB> Stream<M, D> {
             .await?
             .ok_or_else(|| anyhow::Error::msg("unknown twin destination"))?;
 
-        if let Some(relays) = &envelope.relays {
-            let mut twin = self
-                .twins
-                .get_twin(envelope.source.twin)
-                .await?
-                .ok_or_else(|| anyhow::Error::msg("unknown twin source"))?;
-
-            let envelope_relays = match RelayDomains::from_str(relays) {
-                Ok(r) => r,
-                Err(_) => anyhow::bail!("invalid relays"),
-            };
-            if let Some(twin_relays) = twin.relay {
-                if twin_relays != envelope_relays {
-                    twin.relay = Some(envelope_relays);
-                    self.twins.set_twin(twin.clone()).await?;
-                }
-            }
-        }
+        update_cache_relays(&envelope, &self.twins).await?;
 
         if !twin
             .relay
